requirements-dev.txt
asgiref==3.2.10
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
python-decouple==3.3
pytz==2020.1
sqlparse==0.3.1
static3==0.7.0
